Extract output from SimulateCommand to base class
